support adamw for torch_dipu and torch_npu
